move github library functions to new github.rb file
